docker-llm-proxy/training/tf_qwen3_vl/requirements.txt
tensorflow==2.16.1
numpy==1.26.4
Pillow==10.4.0
PyYAML==6.0.2
tqdm==4.66.5
transformers==4.46.3
fastapi==0.115.2
uvicorn==0.32.0
